pyproject.toml
[project]
name = "flagora-backend"
version = "0.1.0"
description = "Backend for the Flagora app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "django-anymail[brevo]>=13.0",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.1",
    "dotenv>=0.9.9",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "requests>=2.32.3",
]
